arm: fsl: csu: Reduce size of ns_dev
None of the values in this struct are larger than 256, so we can reduce
the members to u8s. This saves around 1K.

+++ include/fsl_csu.h
@@ -24,8 +24,8 @@ enum csu_cslx_access {
 };
 
 struct csu_ns_dev {
-	unsigned long ind;
-	uint32_t val;
+	u8 ind;
+	u8 val;
 };
 
 void enable_layerscape_ns_access(void);
